webservice/requirements.txt
pydantic
fastapi
mlflow
pandas-gbq
uvicorn
python-dotenv